Rework internal unit tests
Put all the internal unit tests in a separate file.
The tests suite run (via `ndpiReader`):
* if the option `--run-tests` is used,
* always if the library is compiled with `-DNDPI_EXTENDED_SANITY_CHECKS`
  (useful in CI)

Extend tests using part of #3146.

We keep `tests/unit/unit.c` for serialization tests, like now, because
that is the only code that depends on the external library libjson-c
